# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API de gerenciamento de livros da Biblioteca.",
        "title": "API da Biblioteca",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/categories": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON CategoryRequest e salva a categoria no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Cria um nova categoria",
                "parameters": [
                    {
                        "description": "Dados da Nova categoria a ser criado",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryRequest"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON CategoryRequest e atualiza a categoria no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Atualiza uma categoria",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id da categoria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da nova categoria a ser atualizada",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria atualizada com sucesso"
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui uma categoria específica do banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Exclui uma categoria pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da categoria a ser excluída",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nenhum Conteúdo"
                    },
                    "400": {
                        "description": "Requisição Inválida (ID com formato incorreto)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
        "/public/api/categories": {
            "get": {
                "description": "Retorna uma lista de categorias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Listar categorias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/categories.CategoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/public/api/categories/{id}": {
            "get": {
                "description": "Retorna uma categorias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Obter categoria",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id da categoria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "categories.CategoryRequest": {
            "description": "Dados para criar categoria",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Infantil"
                }
            }
        },
        "categories.CategoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "middleware.APIError": {
            "description": "Erro padronizado de resposta.",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}